fix: read_css_overrides: improve error handling
from #241

Handle file not found separately. This is not really an error situation,
the user just didn't provide any override file. Keep a message for the
time being so users are aware that the feature actually exists, but
make it (hopefully) clearer.

Also include actual error with `read_to_string` failures further down.